Added exported .sql file to seed appointments
This file was dumped from MySQL using phpMyAdmin on XAMPP.  This file
can be imported into another MySql database to generate the initial
appointment data and create the required tables for the Laravel 5
Ultimate Software API challenge.  Make sure to modify the .env file to
point to the right DB and have the correct DB Username and password.